requirements.txt
# 模式识别与机器学习 - 课程大作业依赖包

# 核心科学计算
numpy>=1.19.0
scipy>=1.5.0

# 机器学习
scikit-learn>=0.24.0

# 深度学习(用于加载.pt特征文件)
torch>=1.8.0

# 图像处理
Pillow>=8.0.0

# 可选: 数据可视化
matplotlib>=3.3.0
